+++ app/src/main/java/flab/eryuksa/todocompose/MainActivity.kt
@@ -2,7 +2,9 @@ package flab.eryuksa.todocompose
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
