+++ app/src/main/java/flab/eryuksa/todocompose/TodoApplication.kt
@@ -1,0 +1,7 @@
+package flab.eryuksa.todocompose
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class TodoApplication : Application()
