+++ app/src/main/java/flab/eryuksa/todocompose/data/model/Task.kt
@@ -1,0 +1,10 @@
+package flab.eryuksa.todocompose.data.model
+
+import java.util.UUID
+
+data class Task(
+    val title: String,
+    val details: String,
+    val isDone: Boolean,
+    val id: String = UUID.randomUUID().toString()
+)
