+++ app/src/main/java/flab/eryuksa/todocompose/data/repository/addtodo/AddTodoRepository.kt
@@ -1,0 +1,6 @@
+package flab.eryuksa.todocompose.data.repository.addtodo
+
+interface AddTodoRepository {
+
+    fun addTodo(title: String, details: String)
+}
