+++ app/src/main/java/flab/eryuksa/todocompose/presentation/addtodo/viewmodel/input/AddTodoInput.kt
@@ -1,0 +1,9 @@
+package flab.eryuksa.todocompose.presentation.addtodo.viewmodel.input
+
+interface AddTodoInput {
+
+    fun updateTitle(newTitle: String)
+    fun updateDetails(newDetails: String)
+    fun addTodo()
+    fun dismissScreen()
+}
