+++ app/src/main/java/flab/eryuksa/todocompose/presentation/addtodo/viewmodel/output/AddTodoEffect.kt
@@ -1,0 +1,6 @@
+package flab.eryuksa.todocompose.presentation.addtodo.viewmodel.output
+
+sealed interface AddTodoEffect {
+
+    object DismissAddTodoScreen : AddTodoEffect
+}
