+++ app/src/main/java/flab/eryuksa/todocompose/presentation/addtodo/viewmodel/output/AddTodoOutput.kt
@@ -1,0 +1,10 @@
+package flab.eryuksa.todocompose.presentation.addtodo.viewmodel.output
+
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.StateFlow
+
+interface AddTodoOutput {
+
+    val uiState: StateFlow<AddTodoState>
+    val uiEffect: SharedFlow<AddTodoEffect>
+}
