+++ app/src/main/java/flab/eryuksa/todocompose/presentation/addtodo/viewmodel/output/AddTodoState.kt
@@ -1,0 +1,10 @@
+package flab.eryuksa.todocompose.presentation.addtodo.viewmodel.output
+
+data class AddTodoState(
+    val title: String,
+    val details: String
+) {
+    companion object {
+        val EMPTY = AddTodoState("", "")
+    }
+}
