+++ app/src/main/java/flab/eryuksa/todocompose/presentation/tasks/viewmodel/input/TasksInput.kt
@@ -1,0 +1,9 @@
+package flab.eryuksa.todocompose.presentation.tasks.viewmodel.input
+
+import flab.eryuksa.todocompose.data.model.Task
+
+interface TasksInput {
+
+    fun showAddTaskScreen()
+    fun changeCheckedState(task: Task)
+}
