+++ app/src/main/java/flab/eryuksa/todocompose/presentation/tasks/viewmodel/output/TasksEffect.kt
@@ -1,0 +1,6 @@
+package flab.eryuksa.todocompose.presentation.tasks.viewmodel.output
+
+sealed interface TasksEffect {
+
+    object ShowAddTodoScreen : TasksEffect
+}
