+++ app/src/main/java/flab/eryuksa/todocompose/presentation/tasks/viewmodel/output/TasksOutput.kt
@@ -1,0 +1,10 @@
+package flab.eryuksa.todocompose.presentation.tasks.viewmodel.output
+
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.StateFlow
+
+interface TasksOutput {
+
+    val uiState: StateFlow<TasksState>
+    val uiEffect: SharedFlow<TasksEffect>
+}
