+++ app/src/main/java/flab/eryuksa/todocompose/presentation/theme/Constants.kt
@@ -1,0 +1,3 @@
+package flab.eryuksa.todocompose.presentation.theme
+
+const val DIALOG_HEIGHT_FRACTION = 0.5
