+++ app/src/main/java/flab/eryuksa/todocompose/presentation/theme/Padding.kt
@@ -1,0 +1,11 @@
+package flab.eryuksa.todocompose.presentation.theme
+
+import androidx.compose.ui.unit.dp
+
+object Padding {
+    val NONE = 0.dp
+    val SMALL = 4.dp
+    val MEDIUM = 8.dp
+    val LARGE = 16.dp
+    val XLARGE = 24.dp
+}
