--- app/src/main/java/flab/eryuksa/todocompose/ui/addtodo/viewmodel/input/AddTodoInput.kt
@@ -1,9 +1,0 @@
-package flab.eryuksa.todocompose.ui.addtodo.viewmodel.input
-
-interface AddTodoInput {
-
-    fun updateTitle(newTitle: String)
-    fun updateDetails(newDetails: String)
-    fun addTodo()
-    fun dismissScreen()
-}
