--- app/src/main/java/flab/eryuksa/todocompose/ui/addtodo/viewmodel/output/AddTodoEffect.kt
@@ -1,6 +1,0 @@
-package flab.eryuksa.todocompose.ui.addtodo.viewmodel.output
-
-sealed interface AddTodoEffect {
-
-    object DismissAddTodoScreen : AddTodoEffect
-}
