--- app/src/main/java/flab/eryuksa/todocompose/ui/addtodo/viewmodel/output/AddTodoOutput.kt
@@ -1,10 +1,0 @@
-package flab.eryuksa.todocompose.ui.addtodo.viewmodel.output
-
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.StateFlow
-
-interface AddTodoOutput {
-
-    val uiState: StateFlow<AddTodoState>
-    val uiEffect: SharedFlow<AddTodoEffect>
-}
