--- app/src/main/java/flab/eryuksa/todocompose/ui/addtodo/viewmodel/output/AddTodoState.kt
@@ -1,10 +1,0 @@
-package flab.eryuksa.todocompose.ui.addtodo.viewmodel.output
-
-data class AddTodoState(
-    val title: String,
-    val details: String
-) {
-    companion object {
-        val EMPTY = AddTodoState("", "")
-    }
-}
