--- app/src/main/java/flab/eryuksa/todocompose/ui/tasks/model/Task.kt
@@ -1,3 +1,0 @@
-package flab.eryuksa.todocompose.ui.tasks.model
-
-data class Task(val title: String, val details: String, val isDone: Boolean)
