--- app/src/main/java/flab/eryuksa/todocompose/ui/tasks/viewmodel/input/AddTodoJob.kt
@@ -1,6 +1,0 @@
-package flab.eryuksa.todocompose.ui.tasks.viewmodel.input
-
-interface AddTodoJob {
-
-    fun addTodo(title: String, details: String)
-}
