--- app/src/main/java/flab/eryuksa/todocompose/ui/tasks/viewmodel/input/TasksInput.kt
@@ -1,9 +1,0 @@
-package flab.eryuksa.todocompose.ui.tasks.viewmodel.input
-
-import flab.eryuksa.todocompose.ui.tasks.model.Task
-
-interface TasksInput {
-
-    fun showAddTaskScreen()
-    fun changeCheckState(task: Task)
-}
