--- app/src/main/java/flab/eryuksa/todocompose/ui/tasks/viewmodel/output/TasksEffect.kt
@@ -1,6 +1,0 @@
-package flab.eryuksa.todocompose.ui.tasks.viewmodel.output
-
-sealed interface TasksEffect {
-
-    object ShowAddTodoScreen : TasksEffect
-}
