--- app/src/main/java/flab/eryuksa/todocompose/ui/tasks/viewmodel/output/TasksOutput.kt
@@ -1,10 +1,0 @@
-package flab.eryuksa.todocompose.ui.tasks.viewmodel.output
-
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.StateFlow
-
-interface TasksOutput {
-
-    val uiState: StateFlow<TasksState>
-    val uiEffect: SharedFlow<TasksEffect>
-}
