--- app/src/main/java/flab/eryuksa/todocompose/ui/theme/Constants.kt
@@ -1,3 +1,0 @@
-package flab.eryuksa.todocompose.ui.theme
-
-const val DIALOG_HEIGHT_FRACTION = 0.5
