--- app/src/main/java/flab/eryuksa/todocompose/ui/theme/Padding.kt
@@ -1,11 +1,0 @@
-package flab.eryuksa.todocompose.ui.theme
-
-import androidx.compose.ui.unit.dp
-
-object Padding {
-    val NONE = 0.dp
-    val SMALL = 4.dp
-    val MEDIUM = 8.dp
-    val LARGE = 16.dp
-    val XLARGE = 24.dp
-}
